# ARROW-8996: [C++] AVX2/AVX512 runtime support for aggregate sum kernel