… cpp/src/arrow/CMakeLists.txt
… cpp/src/arrow/compute/kernels/aggregate_test.cc
… cpp/src/arrow/compute/registry_internal.h
… cpp/src/arrow/util/cpu_info.cc
… cpp/src/arrow/util/cpu_info.h
… cpp/src/arrow/util/simd.h