1. Add AVX2/AVX512 path.
2. Add build support to append the compiler flag for SIMD code file.
3. Register the SIMD version at runtime as the CPU feature, also add avx2/avx512 detect.
4. Also add a test case to cover the overflow on the data types which need a right shift
    on the Accumulator type.

Signed-off-by: Frank Du <frank.du@intel.com>