@jianxind 2020-06-01
@wesm Can you help to review the approach for runtime kernel path?

For the build part, I introduced some compiler option(GCC push_options or clang attribute push) to help build the SIMD code without a common SIMD flag defined in CXX flags.

For the SSE implementation, currently only dense part is implemented with intrinsic, I need more work for how to remove the invalid values.