@wesm 2020-06-02
I'll try to review this in the next couple days