@pitrou 2020-06-04
This looks generally quite complicated. If we need 500 additional lines of code to micro-optimize the Sum kernel for a single SIMD instruction set (nevermind that we may also want versions for AVX2, Neon, SVE, and whatnot), things will quickly get out of hand.

If we want to go the way of per-kernel SIMD optimizations, it may be useful to investigate SIMD helper libraries (such as [libsimdpp](https://github.com/p12tic/libsimdpp), [xsimd](https://xsimd.readthedocs.io/en/latest/)...).