@fsaintjacques 2020-06-04
@ursabot benchmark --suite-filter=arrow-compute-aggregate-benchmark --cc=clang --cxx=clang++